Moved async version of json wallet into its own file
